third_party/WebKit/LayoutTests/css3/escape-dom-api-expected.txt

Test window.CSS.escape()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS CSS.escape('\0') is "ï¿½"
PASS CSS.escape('a\0') is "aï¿½"
PASS CSS.escape('\0b') is "ï¿½b"
PASS CSS.escape('a\0b') is "aï¿½b"
PASS CSS.escape('ï¿½') is "ï¿½"
PASS CSS.escape('aï¿½') is "aï¿½"
PASS CSS.escape('ï¿½b') is "ï¿½b"
PASS CSS.escape('aï¿½b') is "aï¿½b"
PASS CSS.escape() threw exception TypeError: Failed to execute 'escape' on 'CSS': 1 argument required, but only 0 present..
PASS CSS.escape(true) is "true"
PASS CSS.escape(false) is "false"
PASS CSS.escape(null) is "null"
PASS CSS.escape('') is ""
PASS CSS.escape('') is "\\1 \\2 \\1e \\1f "
PASS CSS.escape('0a') is "\\30 a"
PASS CSS.escape('1a') is "\\31 a"
PASS CSS.escape('2a') is "\\32 a"
PASS CSS.escape('3a') is "\\33 a"
PASS CSS.escape('4a') is "\\34 a"
PASS CSS.escape('5a') is "\\35 a"
PASS CSS.escape('6a') is "\\36 a"
PASS CSS.escape('7a') is "\\37 a"
PASS CSS.escape('8a') is "\\38 a"
PASS CSS.escape('9a') is "\\39 a"
PASS CSS.escape('a0b') is "a0b"
PASS CSS.escape('a1b') is "a1b"
PASS CSS.escape('a2b') is "a2b"
PASS CSS.escape('a3b') is "a3b"
PASS CSS.escape('a4b') is "a4b"
PASS CSS.escape('a5b') is "a5b"
PASS CSS.escape('a6b') is "a6b"
PASS CSS.escape('a7b') is "a7b"
PASS CSS.escape('a8b') is "a8b"
PASS CSS.escape('a9b') is "a9b"
PASS CSS.escape('-0a') is "-\\30 a"
PASS CSS.escape('-1a') is "-\\31 a"
PASS CSS.escape('-2a') is "-\\32 a"
PASS CSS.escape('-3a') is "-\\33 a"
PASS CSS.escape('-4a') is "-\\34 a"
PASS CSS.escape('-5a') is "-\\35 a"
PASS CSS.escape('-6a') is "-\\36 a"
PASS CSS.escape('-7a') is "-\\37 a"
PASS CSS.escape('-8a') is "-\\38 a"
PASS CSS.escape('-9a') is "-\\39 a"
PASS CSS.escape('-') is "\\-"
PASS CSS.escape('-a') is "-a"
PASS CSS.escape('--') is "--"
PASS CSS.escape('--a') is "--a"
PASS CSS.escape('Â€-_Â©') is "Â€-_Â©"
PASS CSS.escape('Â€ÂÂ‚ÂƒÂ„Â…Â†Â‡ÂˆÂ‰ÂŠÂ‹ÂŒÂÂÂÂÂ‘Â’Â“Â”Â•Â–Â—Â˜Â™ÂšÂ›ÂœÂÂÂŸ') is "\\7f Â€ÂÂ‚ÂƒÂ„Â…Â†Â‡ÂˆÂ‰ÂŠÂ‹ÂŒÂÂÂÂÂ‘Â’Â“Â”Â•Â–Â—Â˜Â™ÂšÂ›ÂœÂÂÂŸ"
PASS CSS.escape('Â Â¡Â¢') is "Â Â¡Â¢"
PASS CSS.escape('a0123456789b') is "a0123456789b"
PASS CSS.escape('abcdefghijklmnopqrstuvwxyz') is "abcdefghijklmnopqrstuvwxyz"
PASS CSS.escape('ABCDEFGHIJKLMNOPQRSTUVWXYZ') is "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
PASS CSS.escape(' !xy') is "\\ \\!xy"
PASS CSS.escape('ğŒ†') is "ğŒ†"
PASS CSS.escape('í¼†') === 'í¼†' is true
PASS CSS.escape('í ´') === 'í ´' is true
PASS successfullyParsed is true

TEST COMPLETE
